backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
ffmpeg-python==0.2.0
imagehash==4.3.1
Pillow==10.2.0
google-generativeai==0.3.2
python-multipart==0.0.9
python-dotenv==1.0.1
google-cloud-pubsub==2.19.0
google-cloud-storage==2.14.0
google-cloud-kms==2.21.3
google-cloud-aiplatform==1.38.1
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9
pinecone-client==3.2.2
redis==5.0.1
